perf: move to bytes instead of words
The code was using words instead of bytes in a lot of places, especially
places where coordinates for squares and offsets were stored. This
implies that a chess board can be larger than 255 coordinates in each
direction, which... isn't possible

This change hasn't had any observable performance benefits yet
unfortunately, at least in terms of running speed, but hopefully the
reduced memory usage will be benefitial

+++ src/board/castling/implementation.rs
@@ -254,7 +254,7 @@ fn calculate_index(side: Side, direction: CastlingDirection) -> usize {
 }
 
 /// Get rank where the pieces of the given `side` start
-fn get_starting_rank(side: Side) -> usize {
+fn get_starting_rank(side: Side) -> u8 {
     match side {
         Side::White => 0,
         Side::Black => 7,
@@ -641,7 +641,7 @@ mod test {
             use super::*;
 
             fn assert_indices_and_files_for_board(
-                indices_and_files: &[(usize, usize)],
+                indices_and_files: &[(usize, u8)],
                 boards: &[Board],
             ) {
                 for &(index, file) in indices_and_files {
